change double to decimal
git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@563 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.Common/DataTransferObjects/CaseLoanDTO.cs
@@ -42,7 +42,7 @@ namespace HPF.FutureState.Common.DataTransferObjects
         public decimal OrigLoanAmt { get; set; }
 
         [NotNullValidator(Ruleset = "Complete", MessageTemplate = "Required!")]
-        public double InterestRate { get; set; }
+        public decimal InterestRate { get; set; }
 
         public string OriginatingLenderName { get; set; }
